defined class Rules; code refactoring
