+++ src/collective/ifttt/profiles/default/actions.xml
@@ -5,7 +5,7 @@
   <object name="Add IFTTT Content Trigger" meta_type="CMF Action" i18n:domain="collective.ifttt">
    <property name="title" i18n:translate="">Add IFTTT Content Trigger</property>
    <property name="description" i18n:translate="">Add new IFTTT Content Trigger</property>
-   <property name="url_expr">string:$object_url/@@add_ifttt_trigger</property>
+   <property name="url_expr">string:$object_url/@@ifttt_content_trigger</property>
    <property name="icon_expr"></property>
    <property name="available_expr">python:path('object/@@checkifttt').check_iftttconfig()</property>
    <property name="permissions">
