+++ src/collective/ifttt/upgrades.zcml
@@ -4,7 +4,7 @@
 
   <genericsetup:upgradeStep
     source="*"
-    destination="1005"
+    destination="1007"
     title="Reload GS profiles"
     description=""
     profile="collective.ifttt:default"
